+++ internal/provider/provider.go
@@ -2,8 +2,9 @@ package provider
 
 import (
 	"context"
-	"fmt"
-	"github.com/atlassian/terraform-provider-atlassian-operations/internal/httpClient"
+	"github.com/atlassian/terraform-provider-atlassian-operations/internal/dto"
+	"github.com/atlassian/terraform-provider-atlassian-operations/internal/provider/dataModels"
+	"github.com/atlassian/terraform-provider-atlassian-operations/internal/provider/schemaAttributes"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/provider"
@@ -14,23 +15,6 @@ import (
 	"os"
 	"time"
 )
-
-// jsmopsProviderModel maps provider schema data to a Go type.
-type jsmopsProviderModel struct {
-	CloudId         types.String `tfsdk:"cloud_id"`
-	DomainName      types.String `tfsdk:"domain_name"`
-	EmailAddress    types.String `tfsdk:"email_address"`
-	Token           types.String `tfsdk:"token"`
-	ApiRetryCount   types.Int32  `tfsdk:"api_retry_count"`
-	ApiRetryWait    types.Int32  `tfsdk:"api_retry_wait"`
-	ApiRetryWaitMax types.Int32  `tfsdk:"api_retry_wait_max"`
-}
-
-type JsmOpsClient struct {
-	OpsClient  *httpClient.HttpClient
-	TeamClient *httpClient.HttpClient
-	UserClient *httpClient.HttpClient
-}
 
 // Ensure the implementation satisfies the expected interfaces.
 var (
@@ -63,39 +47,16 @@ func (p *jsmopsProvider) Metadata(_ context.Context, _ provider.MetadataRequest,
 // Schema defines the provider-level schema for configuration data.
 func (p *jsmopsProvider) Schema(_ context.Context, _ provider.SchemaRequest, resp *provider.SchemaResponse) {
 	resp.Schema = schema.Schema{
-		Attributes: map[string]schema.Attribute{
-			"cloud_id": schema.StringAttribute{
-				Required: true,
-			},
-			"domain_name": schema.StringAttribute{
-				Required: true,
-			},
-			"email_address": schema.StringAttribute{
-				Required: true,
-			},
-			"token": schema.StringAttribute{
-				Required:  true,
-				Sensitive: true,
-			},
-			"api_retry_count": schema.Int32Attribute{
-				Optional: true,
-			},
-			"api_retry_wait": schema.Int32Attribute{
-				Optional: true,
-			},
-			"api_retry_wait_max": schema.Int32Attribute{
-				Optional: true,
-			},
-		},
+		Attributes: schemaAttributes.ProviderAttributes,
 	}
 }
 
-// Configure prepares a atlassian-operations API client for data sources and resources.
+// Configure prepares a atlassian-operations API clientConfiguration for data sources and resources.
 func (p *jsmopsProvider) Configure(ctx context.Context, req provider.ConfigureRequest, resp *provider.ConfigureResponse) {
 	tflog.Info(ctx, "Configuring atlassian-operations provider")
 
 	// Retrieve provider data from configuration
-	var config jsmopsProviderModel
+	var config dataModels.JsmopsProviderTfModel
 	diags := req.Config.Get(ctx, &config)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
@@ -109,7 +70,7 @@ func (p *jsmopsProvider) Configure(ctx context.Context, req provider.ConfigureRe
 		resp.Diagnostics.AddAttributeError(
 			path.Root("cloud_id"),
 			"Unknown cloud instance ID",
-			"The provider cannot create the atlassian-operations API client as there is an unknown configuration value for the cloudId. "+
+			"The provider cannot create the atlassian-operations API clientConfiguration as there is an unknown configuration value for the cloudId. "+
 				"Either target apply the source of the value first, set the value statically in the configuration.",
 		)
 	}
@@ -118,7 +79,7 @@ func (p *jsmopsProvider) Configure(ctx context.Context, req provider.ConfigureRe
 		resp.Diagnostics.AddAttributeError(
 			path.Root("domain_name"),
 			"Unknown domain name",
-			"The provider cannot create the atlassian-operations API client as there is an unknown configuration value for the domain_name. "+
+			"The provider cannot create the atlassian-operations API clientConfiguration as there is an unknown configuration value for the domain_name. "+
 				"Either target apply the source of the value first, set the value statically in the configuration.",
 		)
 	}
@@ -127,7 +88,7 @@ func (p *jsmopsProvider) Configure(ctx context.Context, req provider.ConfigureRe
 		resp.Diagnostics.AddAttributeError(
 			path.Root("email_address"),
 			"Unknown atlassian-operations API EmailAddress",
-			"The provider cannot create the atlassian-operations API client as there is an unknown configuration value for the atlassian-operations API email_address. "+
+			"The provider cannot create the atlassian-operations API clientConfiguration as there is an unknown configuration value for the atlassian-operations API email_address. "+
 				"Either target apply the source of the value first, set the value statically in the configuration.",
 		)
 	}
@@ -136,7 +97,7 @@ func (p *jsmopsProvider) Configure(ctx context.Context, req provider.ConfigureRe
 		resp.Diagnostics.AddAttributeError(
 			path.Root("token"),
 			"Unknown atlassian-operations API Token",
-			"The provider cannot create the atlassian-operations API client as there is an unknown configuration value for the atlassian-operations API token. "+
+			"The provider cannot create the atlassian-operations API clientConfiguration as there is an unknown configuration value for the atlassian-operations API token. "+
 				"Either target apply the source of the value first, set the value statically in the configuration.",
 		)
 	}
@@ -161,7 +122,7 @@ func (p *jsmopsProvider) Configure(ctx context.Context, req provider.ConfigureRe
 
 	cloudId := os.Getenv("ATLASSIAN_OPS_CLOUD_ID")
 	domainName := os.Getenv("ATLASSIAN_OPS_DOMAIN_NAME")
-	email_address := os.Getenv("ATLASSIAN_OPS_API_USERNAME")
+	emailAddress := os.Getenv("ATLASSIAN_OPS_API_USERNAME")
 	token := os.Getenv("ATLASSIAN_OPS_API_TOKEN")
 
 	if cloudId == "" {
@@ -169,7 +130,7 @@ func (p *jsmopsProvider) Configure(ctx context.Context, req provider.ConfigureRe
 			resp.Diagnostics.AddAttributeError(
 				path.Root("cloud_id"),
 				"Invalid cloud instance ID",
-				"The provider cannot create the atlassian-operations API client as there is a null / an empty configuration value for the cloudId.",
+				"The provider cannot create the atlassian-operations API clientConfiguration as there is a null / an empty configuration value for the cloudId.",
 			)
 		} else {
 			cloudId = config.CloudId.ValueString()
@@ -181,22 +142,22 @@ func (p *jsmopsProvider) Configure(ctx context.Context, req provider.ConfigureRe
 			resp.Diagnostics.AddAttributeError(
 				path.Root("domain_name"),
 				"Unknown domain name",
-				"The provider cannot create the atlassian-operations API client as there is an unknown configuration value for the domain_name. ",
+				"The provider cannot create the atlassian-operations API clientConfiguration as there is an unknown configuration value for the domain_name. ",
 			)
 		} else {
 			domainName = config.DomainName.ValueString()
 		}
 	}
 
-	if email_address == "" {
+	if emailAddress == "" {
 		if config.EmailAddress.IsNull() {
 			resp.Diagnostics.AddAttributeError(
 				path.Root("email_address"),
 				"Unknown atlassian-operations API EmailAddress",
-				"The provider cannot create the atlassian-operations API client as there is an unknown configuration value for the atlassian-operations API email_address. ",
+				"The provider cannot create the atlassian-operations API clientConfiguration as there is an unknown configuration value for the atlassian-operations API email_address. ",
 			)
 		} else {
-			email_address = config.EmailAddress.ValueString()
+			emailAddress = config.EmailAddress.ValueString()
 		}
 	}
 
@@ -205,7 +166,7 @@ func (p *jsmopsProvider) Configure(ctx context.Context, req provider.ConfigureRe
 			resp.Diagnostics.AddAttributeError(
 				path.Root("token"),
 				"Unknown atlassian-operations API Token",
-				"The provider cannot create the atlassian-operations API client as there is an unknown configuration value for the atlassian-operations API token. ",
+				"The provider cannot create the atlassian-operations API clientConfiguration as there is an unknown configuration value for the atlassian-operations API token. ",
 			)
 		} else {
 			token = config.Token.ValueString()
@@ -214,47 +175,30 @@ func (p *jsmopsProvider) Configure(ctx context.Context, req provider.ConfigureRe
 
 	ctx = tflog.SetField(ctx, "atlassian-operations_cloud_id", cloudId)
 	ctx = tflog.SetField(ctx, "atlassian-operations_domain_name", domainName)
-	ctx = tflog.SetField(ctx, "atlassian-operations_email_address", email_address)
+	ctx = tflog.SetField(ctx, "atlassian-operations_email_address", emailAddress)
 	ctx = tflog.SetField(ctx, "atlassian-operations_token", token)
 	ctx = tflog.MaskFieldValuesWithFieldKeys(ctx, "atlassian-operations_token")
 
-	tflog.Debug(ctx, "Creating atlassian-operations client")
+	tflog.Debug(ctx, "Creating atlassian-operations clientConfiguration")
 
-	// Create a new atlassian-operations client using the configuration values
-	client := &JsmOpsClient{
-		OpsClient: httpClient.
-			NewHttpClient().
-			SetRetryCount(int(config.ApiRetryCount.ValueInt32())).
-			SetRetryWaitTime(time.Duration(config.ApiRetryWait.ValueInt32())*time.Second).
-			SetRetryMaxWaitTime(time.Duration(config.ApiRetryWaitMax.ValueInt32())*time.Second).
-			SetDefaultBasicAuth(email_address, token).
-			SetBaseUrl(fmt.Sprintf("https://api.atlassian.com/jsm/ops/api/%s", cloudId)),
-		TeamClient: httpClient.
-			NewHttpClient().
-			SetRetryCount(int(config.ApiRetryCount.ValueInt32())).
-			SetRetryWaitTime(time.Duration(config.ApiRetryWait.ValueInt32())*time.Second).
-			SetRetryMaxWaitTime(time.Duration(config.ApiRetryWaitMax.ValueInt32())*time.Second).
-			SetDefaultBasicAuth(email_address, token).
-			SetBaseUrl(fmt.Sprintf("https://%s/gateway/api/public/teams/v1/org/", domainName)),
-		UserClient: httpClient.
-			NewHttpClient().
-			SetRetryCount(int(config.ApiRetryCount.ValueInt32())).
-			SetRetryWaitTime(time.Duration(config.ApiRetryWait.ValueInt32())*time.Second).
-			SetRetryMaxWaitTime(time.Duration(config.ApiRetryWaitMax.ValueInt32())*time.Second).
-			SetDefaultBasicAuth(email_address, token).
-			SetBaseUrl(fmt.Sprintf("https://%s/rest/api/3/user/", domainName)),
-	}
+	// Create a new atlassian-operations clientConfiguration using the configuration values
+	client := dto.NewJsmopsProviderModel(
+		cloudId,
+		domainName,
+		emailAddress,
+		token,
+		int(config.ApiRetryCount.ValueInt32()),
+		time.Duration(config.ApiRetryWait.ValueInt32())*time.Second,
+		time.Duration(config.ApiRetryWaitMax.ValueInt32())*time.Second,
+		isStaging,
+	)
 
-	if isStaging {
-		client.OpsClient = client.OpsClient.SetBaseUrl(fmt.Sprintf("https://api.stg.atlassian.com/jsm/ops/api/%s", cloudId))
-	}
-
-	// Make the atlassian-operations client available during DataSource and Resource
+	// Make the atlassian-operations clientConfiguration available during DataSource and Resource
 	// type Configure methods.
 	resp.DataSourceData = client
 	resp.ResourceData = client
 
-	tflog.Info(ctx, "Configured atlassian-operations client", map[string]any{"success": true})
+	tflog.Info(ctx, "Configured atlassian-operations clientConfiguration", map[string]any{"success": true})
 }
 
 // DataSources defines the data sources implemented in the provider.