+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v1.0.1
+
+### FIXES:
+
+- Fixed a race condition issue due to consequent requests sent by the same HTTP client effecting each other.
+- Fixed typos in schedule and schedule rotation import scripts
+
 ## v1.0.0
 
 ### FEATURES: